+++ api/Models/Payments/PaymentResult.cs
@@ -7,9 +7,9 @@ public class PaymentResult
     public DateTime? ApprovalDate { get; set; } // Data da aprovação (caso ocorra)
     public DateTime? DisapprovalDate { get; set; } // Data de reprovação (caso ocorra)
     public bool Confirmation { get; set; } // Confirmação da adquirente
-    public float GrossValue { get; set; } // Valor bruto da transação
-    public float? NetValue { get; set; } // Valor líquido da transação (descontado taxa)
-    public float? FlatRate { get; set; } // Taxa fixa cobrada
+    public decimal GrossValue { get; set; } // Valor bruto da transação
+    public decimal? NetValue { get; set; } // Valor líquido da transação (descontado taxa)
+    public decimal? FlatRate { get; set; } // Taxa fixa cobrada
     public string CardNumber { get; set; } // 4 ultimos digitos do cartão
     public ICollection<InstallmentResult> Installments { get; set; } // Lista das parcelas
 }
